requirements.txt
# Options Day-Trading Assistant - Dependencies

# TD Ameritrade API
tda-api==1.3.0

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# Options pricing and Greeks
scipy==1.11.4
py_vollib==1.0.1

# Database
sqlalchemy==2.0.23

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Terminal output
rich==13.7.0

# HTTP requests (for API fallback)
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Date/time utilities
python-dateutil==2.8.2
